backend/requirements.txt
djangorestframework==3.14.0
Django==4.2.7
drf-yasg==1.21.7
django-cors-headers==4.3.1